Only create symlinks on unix.

+++ test/tm.test.js
@@ -88,8 +88,11 @@ describe('tm', function() {
     });
 
     it('dirfiles', function(done) {
-        fs.symlinkSync('broken',path.join(__dirname,'fixtures-localsource','broken-symlink'));
-        assert.ok(fs.existsSync(__dirname,'fixtures-localsource','broken-symlink'));
+        var platform = require('os').platform();
+        if (platform === 'linux' || platform === 'darwin') {
+            fs.symlinkSync('broken',path.join(__dirname,'fixtures-localsource','broken-symlink'));
+            assert.ok(fs.existsSync(__dirname,'fixtures-localsource','broken-symlink'));
+        }
         tm.dirfiles(__dirname + '/fixtures-localsource', function(err, files) {
             assert.ifError(err);
             assert.deepEqual([
@@ -101,7 +104,9 @@ describe('tm', function() {
                 'data.yml',
                 'project.yml'
             ], files.map(function(f) { return f.basename }));
-            fs.unlinkSync(path.join(__dirname,'fixtures-localsource','broken-symlink'));
+            if (platform === 'linux' || platform === 'darwin') {
+                fs.unlinkSync(path.join(__dirname,'fixtures-localsource','broken-symlink'));
+            }
             done();
         });
     });
